Unify type declarations in typescript into a single file.
